Fix: Update type variable bug 7

+++ smrt/src/main/java/com/smart/dto/ProduitDTO.java
@@ -2,13 +2,12 @@ package com.smart.dto;
 
 import lombok.Data;
 
-import java.math.BigDecimal;
 
 @Data
 public class ProduitDTO {
     private String id;
     private String nom;
     private String categorie;
-    private BigDecimal poids;
+    private Double poids;
     private Double prix;
 }
